+++ CotEditor/Sources/Settings Window/Syntax Editor/SyntaxHighlightEditView.swift
@@ -32,7 +32,6 @@ struct SyntaxHighlightEditView: View {
     
     
     @Binding var items: [Item]
-    var helpAnchor: String = "syntax_highlight_settings"
     
     @State private var selection: Set<Item.ID> = []
     @State private var sortOrder: [KeyPathComparator<Item>] = []
@@ -131,8 +130,11 @@ struct SyntaxHighlightEditView: View {
                     .monospacedDigit()
                     .foregroundStyle(.secondary)
                     .controlSize(.small)
+            }
+            
+            HStack {
                 Spacer()
-                HelpLink(anchor: self.helpAnchor)
+                HelpLink(anchor: "syntax_highlight_settings")
             }
         }
     }